+++ tests/BaseTests.php
@@ -11,6 +11,7 @@ use Cviebrock\EloquentSluggable\Tests\Models\PostWithCustomSource;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithCustomSuffix;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithEmptySeparator;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithMaxLength;
+use Cviebrock\EloquentSluggable\Tests\Models\PostWithMaxLengthSplitWords;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithMultipleSlugs;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithMultipleSources;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithNoSource;
@@ -201,6 +202,17 @@ class BaseTests extends TestCase
         $post = PostWithMaxLength::create([
             'title' => 'A post with a really long title'
         ]);
+        $this->assertEquals('a-post', $post->slug);
+    }
+
+    /**
+     * Test for max_length option with word splitting.
+     */
+    public function testMaxLengthSplitWords()
+    {
+        $post = PostWithMaxLengthSplitWords::create([
+            'title' => 'A post with a really long title'
+        ]);
         $this->assertEquals('a-post-wit', $post->slug);
     }
 
@@ -211,6 +223,23 @@ class BaseTests extends TestCase
     {
         for ($i = 0; $i < 20; $i++) {
             $post = PostWithMaxLength::create([
+                'title' => 'A post with a really long title'
+            ]);
+            if ($i == 0) {
+                $this->assertEquals('a-post', $post->slug);
+            } elseif ($i < 10) {
+                $this->assertEquals('a-post-' . $i, $post->slug);
+            }
+        }
+    }
+
+    /**
+     * Test for max_length option with increments and word splitting.
+     */
+    public function testMaxLengthSplitWordsWithIncrements()
+    {
+        for ($i = 0; $i < 20; $i++) {
+            $post = PostWithMaxLengthSplitWords::create([
                 'title' => 'A post with a really long title'
             ]);
             if ($i == 0) {
@@ -226,7 +255,7 @@ class BaseTests extends TestCase
      */
     public function testMaxLengthDoesNotEndInSeparator()
     {
-        $post = PostWithMaxLength::create([
+        $post = PostWithMaxLengthSplitWords::create([
             'title' => 'It should work'
         ]);
         $this->assertEquals('it-should', $post->slug);